#10 Add dedicated MidiLaunchpadSClient implementation
