+++ launchpad-s/midi/src/test/kotlin/io/lp4k/launchpad/midi/s/MidiLaunchpadSClientTest.kt
@@ -1,6 +1,5 @@
 /*
  *    Copyright 2020 Vasyl Rudas
- *    Copyright 2015 Olivier Croisier (thecodersbreakfast.net)
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -15,7 +14,7 @@
  *    limitations under the License.
  *
  */
-package io.lp4k.midi
+package io.lp4k.launchpad.midi.s
 
 import io.lp4k.launchpad.api.BackBufferOperation
 import io.lp4k.launchpad.api.Buffer
@@ -36,7 +35,7 @@ import org.mockito.junit.jupiter.MockitoExtension
 import javax.sound.midi.InvalidMidiDataException
 
 @ExtendWith(MockitoExtension::class)
-class MidiLaunchpadClientTest {
+class MidiLaunchpadSClientTest {
 
     private lateinit var launchpadClient: LaunchpadClient
 
@@ -46,7 +45,7 @@ class MidiLaunchpadClientTest {
     @BeforeEach
     fun init() {
         midiProtocolClient = Mockito.mock(MidiProtocolClient::class.java)
-        launchpadClient = MidiLaunchpadClient(midiProtocolClient)
+        launchpadClient = MidiLaunchpadSClient(midiProtocolClient)
     }
 
     /*
